+++ website/src/repl/util.mjs
@@ -9,40 +9,12 @@ import { createClient } from '@supabase/supabase-js';
 import { nanoid } from 'nanoid';
 import { writeText } from '@tauri-apps/api/clipboard';
 import { createContext } from 'react';
-import { $publicPatterns, $featuredPatterns } from '../settings.mjs';
 
 // Create a single supabase client for interacting with your database
 export const supabase = createClient(
   'https://pidxdsxphlhzjnzmifth.supabase.co',
   'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6InBpZHhkc3hwaGxoempuem1pZnRoIiwicm9sZSI6ImFub24iLCJpYXQiOjE2NTYyMzA1NTYsImV4cCI6MTk3MTgwNjU1Nn0.bqlw7802fsWRnqU5BLYtmXk_k-D1VFmbkHMywWc15NM',
 );
-
-export function loadPublicPatterns() {
-  return supabase.from('code').select().eq('public', true).limit(20).order('id', { ascending: false });
-}
-
-export function loadFeaturedPatterns() {
-  return supabase.from('code').select().eq('featured', true).limit(20).order('id', { ascending: false });
-}
-
-async function loadDBPatterns() {
-  try {
-    const { data: publicPatterns } = await loadPublicPatterns();
-    const { data: featuredPatterns } = await loadFeaturedPatterns();
-    const featured = {};
-    const pub = {};
-    publicPatterns?.forEach((data, key) => (pub[data.id ?? key] = data));
-    featuredPatterns?.forEach((data, key) => (featured[data.id ?? key] = data));
-    $publicPatterns.set(pub);
-    $featuredPatterns.set(featured);
-  } catch (err) {
-    console.error('error loading patterns');
-  }
-}
-
-if (typeof window !== 'undefined') {
-  loadDBPatterns();
-}
 
 export async function initCode() {
   // load code from url hash (either short hash from database or decode long hash)
